+++ changelog.rst
@@ -9,7 +9,10 @@ Upcoming
 * Remove `video` and `video_hours_count` tempate tags.
 * Remove flowplayer and all javascript.
 * OPEN SOURCE! WOO!
-* Add sorl-thumbnail to dependancies. (This was already used in the template, just not mentioned.)
+* Replace `sorl-thumbnail` with `django-imagekit`.
+* Make "Chapter.seconds" return an int
+* Add VideoContent
+* Set related_name on Source.video and Chapter.video
 
 v1.0
 ------